% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.causalTree.R
\name{estimate.causalTree}
\alias{estimate.causalTree}
\title{Estimate the causal effects using honest tree model.}
\usage{
estimate.causalTree(object, formula, data, weights, na.action = na.pass)
}
\arguments{
\item{object}{A tree-structured fit object.}

\item{formula}{A regression formula.}

\item{data}{New observations.}

\item{weights}{The treatment status of new observations}
}
\value{
The estimated causal effects of \code{data}.
Notice here when the leaf contains only treated or control cases, the function will trace back to the leaf's parent mnode recursively until the parent can be used to compute causal effect.
}
\description{
Estimate the causal effects using honest tree model.
}

